# openapi-updated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new user",
        "description": "Create a new user account with email, password, and username.\n    \n    - **Email Verification**: If enabled in Supabase, user must verify email before signing in.\n    - **Username**: Must be 3-30 characters, alphanumeric and underscores only.\n    - **Password**: Must be at least 8 characters with uppercase, lowercase, and digit.\n    - **Rate Limit**: 5 requests per minute.",
        "operationId": "sign_up_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or user already exists"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit exceeded",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded"
                }
              }
            }
          },
          "502": {
            "description": "Failed to connect to Supabase"
          }
        }
      }
    },
...
    "/auth/resend-verification": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Resend verification email",
        "description": "Resend the verification email for users who haven't verified their email yet.\n    \n    - **Use Case**: When verification link expires or email was not received.\n    - **Rate Limit**: 3 requests per minute.",
        "operationId": "resend_verification_auth_resend_verification_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Verification email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Could not send verification email"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit exceeded",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user profile",
        "description": "Search for a user by email.\n\nArgs:\n    email: The email address to look up.\n    current_user: The authenticated user making the request.\n    service: The user management service.\n\nReturns:\n    `UserSearchResponse` containing the user's id, email, f_name, and l_name.",
        "operationId": "get_user_profile_users__get",
        "responses": {
          "200": {
            "description": "User found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      }
    },
...
    "/documents/{document_id}/parties": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List document parties",
        "description": "Return the issuer and all whitelisted parties for a document. Accessible by owner, any party, or admin.",
        "operationId": "get_document_parties_documents__document_id__parties_get",
        "security": [
          {
            "Bearer Token": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of parties",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPartyListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — missing or invalid bearer token"
          },
          "404": {
            "description": "Document not found or user has no access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Add a party to a document",
        "description": "Whitelist a user by email to access this document. Only the document owner (or admin) may add parties. Roles: viewer | signer | editor — signer and editor may only be granted to lawyers.",
        "operationId": "add_document_party_documents__document_id__parties_post",
        "security": [
          {
            "Bearer Token": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPartyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Party added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPartyResponse"
                }
              }
            }
          },
          "400": {
            "description": "User does not exist, or role is invalid for this user"
          },
          "401": {
            "description": "Not authenticated — missing or invalid bearer token"
          },
          "403": {
            "description": "Forbidden — only the document owner may add parties"
          },
          "404": {
            "description": "Document not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/parties/{party_user_id}": {
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Remove a party from a document",
        "description": "Revoke a user's access to this document. Only the document owner (or admin) may remove parties.",
        "operationId": "remove_document_party_documents__document_id__parties__party_user_id__delete",
        "security": [
          {
            "Bearer Token": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "party_user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Party User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Party removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemovePartyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — missing or invalid bearer token"
          },
          "403": {
            "description": "Forbidden — only the document owner may remove parties"
          },
          "404": {
            "description": "Document or party not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddPartyRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Email of the user to whitelist."
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Role: viewer | signer | editor",
            "default": "viewer"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "AddPartyRequest",
        "description": "Request to whitelist a user on a document"
      },
...
      "DocumentPartyListResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id",
            "description": "Document identifier."
          },
          "issuer": {
            "$ref": "#/components/schemas/DocumentPartyResponse",
            "description": "The owner/issuer of the document"
          },
          "parties": {
            "items": {
              "$ref": "#/components/schemas/DocumentPartyResponse"
            },
            "type": "array",
            "title": "Parties"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "issuer"
        ],
        "title": "DocumentPartyListResponse",
        "description": "List of parties for a document"
      },
      "DocumentPartyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Party record ID."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "User's email"
          },
          "f_name": {
            "type": "string",
            "title": "F Name",
            "description": "User's first Name"
          },
          "l_name": {
            "type": "string",
            "title": "L Name",
            "description": "User's last name"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Role assigned"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "pending | accepted | declined — the owner is always accepted",
            "default": "accepted"
          },
          "responded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responded At",
            "description": "When this party accepted or declined"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "email",
          "f_name",
          "l_name",
          "role",
          "created_at"
        ],
        "title": "DocumentPartyResponse",
        "description": "A single party record on a document"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "RemovePartyResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "user_id",
          "message"
        ],
        "title": "RemovePartyResponse",
        "description": "Confirmation after removing a party"
      },
...
      "ResendVerificationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email address to resend verification to"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ResendVerificationRequest"
      },
...
      "SignUpRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Valid email for this user"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Valid Password"
          },
          "f_name": {
            "type": "string",
            "title": "F Name",
            "description": "Valid first name"
          },
          "l_name": {
            "type": "string",
            "title": "L Name",
            "description": "Valid last name"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number",
            "description": "Valid Phone number"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "f_name",
          "l_name",
          "phone_number"
        ],
        "title": "SignUpRequest"
      },
      "SignUpResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Message containing what to do next."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "Contains the user id the response"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email of this user/response"
          },
          "requires_email_confirmation": {
            "type": "boolean",
            "title": "Requires Email Confirmation",
            "description": "session will be None until the user verifies their email"
          }
        },
        "type": "object",
        "required": [
          "message",
          "user_id",
          "email",
          "requires_email_confirmation"
        ],
        "title": "SignUpResponse"
      },
...
      "UserProfileResponse": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "description": "User's email address"
          },
          "f_name": {
            "type": "string",
            "title": "F Name",
            "description": "Valid first name"
          },
          "l_name": {
            "type": "string",
            "title": "L Name",
            "description": "Valid last name"
          },
          "avatar": {
            "type": "string",
            "title": "Avatar",
            "description": "User Avatar",
            "default": "icon1"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "User role"
          }
        },
        "type": "object",
        "required": [
          "email",
          "f_name",
          "l_name",
          "role"
        ],
        "title": "UserProfileResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "Bearer Token": {
        "type": "http",
        "description": "Enter your Supabase JWT token",
        "scheme": "bearer"
      }
    }
  }
}